Preserve service sync config for strict validation

+++ config/service_sync.php
@@ -3,9 +3,9 @@
 declare(strict_types=1);
 
 return [
-    'interval_minutes' => (int) env('SERVICE_SYNC_INTERVAL_MINUTES', 5),
-    'batch_limit' => (int) env('SERVICE_SYNC_BATCH_LIMIT', 50),
-    'lease_seconds' => (int) env('SERVICE_SYNC_LEASE_SECONDS', 120),
+    'interval_minutes' => env('SERVICE_SYNC_INTERVAL_MINUTES', 5),
+    'batch_limit' => env('SERVICE_SYNC_BATCH_LIMIT', 50),
+    'lease_seconds' => env('SERVICE_SYNC_LEASE_SECONDS', 120),
     'severity' => [
         'missing_remote' => env('SERVICE_SYNC_SEVERITY_MISSING_REMOTE', 'critical'),
         'expired_local_active_remote' => env('SERVICE_SYNC_SEVERITY_EXPIRED_ACTIVE', 'warning'),
